comments resolved

+++ app/controllers/order_items_controller.rb
@@ -2,9 +2,17 @@ class OrderItemsController < ApplicationController
   load_and_authorize_resource :order
   load_and_authorize_resource :order_item, through: :order
 
+  #GET /order/:order_id/order_items
+  def index
+    @order_items = @order_items.order(created_at: :desc).paginate(page: params[:page], per_page: LISTING_PAGINATION_SIZE)
+    respond_to do |format|
+      format.html
+    end
+  end
+
   #GET /order/:order_id/order_items/new
   def new
-    @items = Item.all
+    @items = Item.active
     respond_to do |format|
       format.js
     end
@@ -12,7 +20,7 @@ class OrderItemsController < ApplicationController
 
   #GET /order/:order_id/order_items/:id/edit
   def edit
-    @items = Item.all
+    @items = Item.active
     respond_to do |format|
       format.html
     end
@@ -20,59 +28,65 @@ class OrderItemsController < ApplicationController
 
   #POST /order/:order_id/order_items
   def create
-    @items = Item.all
+    @items = Item.active
+    create_success = false
     if @order_item.item.quantity >= @order_item.quantity
       OrderItem.transaction do
         if @order_item.save
           @order_item.item.update(quantity: (@order_item.item.quantity - @order_item.quantity))
-          flash[:notice] = 'Order item successfully added.'
-          respond_to do |format|
-            format.js
-          end
-        else
-          flash[:alert] = "Order item could not be added beacuse of #{@order_item.errors.full_messages}."
-          respond_to do |format|
-            format.js
-          end
+          create_success = true
         end
       end
     else
-      flash[:alert] = 'Order item quantity cannot be greater than item stock quantity.'
-      respond_to do |format|
-        format.js
-        end
+      @order_item.errors.add(:quantity, I18n.t(:order_item_quantity_error))
+    end
+    respond_to do |format|
+      format.js do
+       render action: "failure" unless create_success
+      end
     end
   end
 
   #PUT /order/:order_id/order_items/:id
   def update
-    @items = Item.all
-    if check_updated_quantity
-      if @order_item.update(order_item_params)
-        respond_to do |format|
-          flash[:notice] = 'Order item successfully updated.'
-          format.html { redirect_to [@order, @order_item] }
+    @items = Item.active
+    success = false
+    OrderItem.transaction do
+      if @order_item.check_updated_quantity(params[:order_item][:item_id].to_i, params[:order_item][:quantity].to_i)
+        if @order_item.update(order_item_params)
+          success = true
+          flash[:notice] = I18n.t(:order_item_update_success)
+        else
+          flash.now[:alert] = I18n.t(:order_item_update_fail, error: @order_item.errors.full_messages.to_sentence)
         end
       else
-        flash.now[:alert] = 'Order item could not be updated.'
-        respond_to do |format|
-          format.html { render 'edit' }
-        end
+        flash.now[:alert] = I18n.t(:order_item_update_quantity_error)
       end
-    else
-      flash[:alert] = 'Order items could not be updated, not enough items in stock.'
-      respond_to do |format|
-        format.html { render 'edit' }
+    end
+    respond_to do |format|
+      format.html do
+        if success
+          redirect_to [@order, @order_item]
+        else
+          render 'edit'
+        end
       end
     end
   end
 
   #DELETE /order/:order_id/order_items/:id
   def destroy
-    if @order_item.destroy
-      flash[:notice] = 'Order item successfully deleted.'
+    if @order_item.order.status != 'Completed'
+      OrderItem.transaction do
+        if @order_item.destroy
+          flash[:notice] = I18n.t(:order_item_delete_success)
+          @order_item.item.update(quantity: @order_item.item.quantity + @order_item.quantity)
+        else
+          flash[:alert] = I18n.t(:order_item_delete_fail, error: @order_item.errors.full_messages.to_sentence)
+        end
+      end
     else
-      flash[:alert] = 'Order item could not be deleted.'
+      flash[:alert] = I18n.t(:order_item_destroy_status_error)
     end
     respond_to do |format|
       format.html { redirect_to order_order_items_path(@order) }
@@ -84,12 +98,4 @@ class OrderItemsController < ApplicationController
     params.require(:order_item).permit(:item_id, :quantity)
   end
 
-  def check_updated_quantity
-    if (params[:order_item][:quantity]).to_i > @order_item.quantity
-      quantity_diff  = @order_item.item.quantity - (params[:order_item][:quantity].to_i - @order_item.quantity)
-    else
-      quantity_diff  = @order_item.item.quantity + (@order_item.quantity - params[:order_item][:quantity].to_i)
-    end
-    @order_item.item.update(quantity: quantity_diff)
-  end
 end
